==> R/55_MCMC_plot.Rmd <==
---
title: "sample Instrument and Proxy"
author: "Nicholas Nagle"
date: "February 13, 2015"
output: pdf_document
---
```{r }
library(dplyr)
library(tidyr)
library(Matrix)
library(INLA)
library(ggplot2)
library(scales)

# Load the sat data:
load('~/Dropbox/git_root/climate-bayes/data/indiv.station.Rdata')
# Load the tree ring data:
load('~/Dropbox/git_root/climate-bayes/data/itrdb_meta.Rdata')
# Load the spatial matrices that were created by 25_create_spatial_field.R
load('~/Dropbox/git_root/climate-bayes/data/spatial_fields.Rdata')

# Load the MCMC output
load('~/Dropbox/git_root/climate-bayes/mcout.Rdata')

# Determine "burn-in"
burn <- 100
keeps <- round(seq(from=1000, to = length(mc$rho), length=1000))
```

# plot the posterior tree coefficients and tree locations 
```{r}
library("PBSmapping")
xlim=c(-180, -40)
ylim=c(25,75)
worldmap = map_data("world")
names(worldmap) <- c("X","Y","PID","POS","region","subregion")
worldmap = clipPolys(worldmap, xlim=xlim,ylim=ylim, keepExtra=TRUE)

beta_posterior <- colMeans(mc$beta.1[keeps,])
names(beta_posterior) <- species_levels
tree_loc <- tree.meta %>% mutate( beta_1=beta_posterior[as.character(tree.meta$species)] )
plt <- ggplot(data=tree_loc) + geom_point(aes(x=-lon, y=lat, color=beta_1)) + 
	scale_color_gradient2( low=muted("blue"), high=muted("red"), midpoint=0 ) +
coord_map(projection="mollweide") + 
geom_polygon( data=worldmap, aes(X,Y,group=PID), fill=NA, color="grey50" )
print(plt)
ggsave(filename="beta.pdf", plot=plt, width=11, height=8.5)


plt <- ggplot(data=tree_loc) + geom_point(aes(x=-lon, y=lat, color=species_code))


plt.data <- mc$beta.1[keeps,]
plt.data <- as.data.frame(plt.data)
names(plt.data) <- species_levels
plt.data <- gather(plt.data, key=species, value=beta_1)
plt <- plt.data %>% filter(species!="INST") %>% ggplot() + geom_boxplot( aes(x=species, y=beta_1) )
print(plt)
ggsave(filename="beta_boxplots.pdf", plot=plt, width=11, height=8.5)
```


plot the climate field

This is not the best way to do it.  I am just averaging the instrument/tree locations within grid cells for now.  Better would be to create a grid of points and then predict there.

```{r}
# Create a high res grid of points:
pred_locs <- expand.grid( lat = seq(30,80,2), lon = seq(-180,-50,2) )
# overlay that grid on the continent
coordinates(pred_locs) <- ~ lon + lat
proj4string(pred_locs) <- '+proj=longlat'
# convert worldmap into spatial polygons:
attr(worldmap, 'projection') <- 'LL'
world.sp <- maptools::PolySet2SpatialPolygons(worldmap)
#world.sp %over% pred_locs
pred_locs <- pred_locs[which(!is.na(pred_locs %over% world.sp)),]
#plot(pred_locs)
# Create the basis
A.pred <- as.data.frame(pred_locs) %>% select(lon, lat) %>% as.matrix %>%
inla.mesh.map(loc=.,projection='longlat') %>% inla.spde.make.A(mesh=mesh2, loc=.)

# Create the MCMC predictions:
dimnames(mc.alpha) <- list(mc=seq(1,dim(mc.alpha)[1]), ID=seq(1,dim(mc.alpha)[2]), year=years)
dimnames(A.pred) <- list(PID=seq(1, dim(A.pred)[1]), ID=seq(1, dim(A.pred)[2]))
preds <- apply(mc.alpha[keeps,,], c(1,3), function(x) as.vector(A.pred %*% x))
dimnames(preds) <- list(PID=seq(1,dim(A.pred)[1]), mc=seq(1,dim(preds)[2]), year=years)
# Flatten preds
preds <- reshape2::melt(preds)

pred_locs <- as.data.frame(pred_locs) %>% mutate(PID=seq(1,length(pred_locs))) %>%
	mutate(wt=sin((lat+1)/180) - sin((lat-1)/180))

preds <- left_join(preds, pred_locs)
plt <- preds %>% filter(year > 1980 & year <=2000) %>% 
group_by(lon, lat, year) %>% summarize(anomaly=mean(value)) %>% 
ggplot()+geom_tile( aes(x=lon, y=lat, fill=anomaly )) + facet_wrap(~year) +
scale_fill_gradient2(low=muted("blue"), high=muted("red"), midpoint=0, limit=c(-4,4))+
coord_map(projection="mollweide") +
geom_polygon(data=worldmap, aes(X,Y,group=PID), fill=NA, color="grey50")
print(plt)
ggsave(filename="anomaly_ts.pdf", plot=plt, height=8.5, width=11)
```

plot the north american average time series
```{r}
plt <- preds %>% 
group_by(mc, year) %>% summarize(value=sum(value*wt)/sum(wt)) %>%
group_by(year) %>% 
summarize(low=quantile(value, .05), mn=mean(value), high=quantile(value, .95)) %>% 
ggplot() + geom_ribbon(aes(x=year, ymin=low, ymax=high), fill='grey50') +
geom_line(aes(x=year, y=mn)) + labs('y=anomaly')
print(plt)
ggsave(filename="North_America_ts.pdf", plot=plt, height=8.5, width=11)
```

plot regional averages
```{r}

plt <- preds %>%
mutate(lat2 = round(lat/10)*10, lon2 = round(lon/10)*10) %>%
group_by(mc, year, lat2, lon2) %>% summarize(value=sum(value*wt)/sum(wt)) %>%
group_by(lat2, lon2, year) %>%
summarize(low=quantile(value, .05), mn=mean(value), high=quantile(value, .95)) %>% 
mutate(lat3 = factor(lat2, levels=c(80, 70, 60, 50, 40, 30))) %>%
ggplot() + geom_ribbon(aes(x=year, ymin=low, ymax=high), fill='grey50') +
geom_line(aes(x=year, y=mn)) + labs('y=anomaly')+facet_grid(lat3~lon2)+scale_y_continuous(limits=c(-4,4))
print(plt)
ggsave(filename="regional_ts.pdf", plot=plt, height=8.5, width=11)
```

plot species time series:
```{r}

B.year.plt <- bs(x=years, 
                  knots=seq(min(years),max(years), length=k.yr)[-c(1,k.yr)],
                  df=3,
                  Boundary.knots=r, intercept=TRUE)


beta.temp <- sweep(mc$beta.year.1, c(1,2), mc$beta.1, FUN='+')
temp <- apply(beta.temp[keeps,,], c(1,2), function(x) as.vector(B.year.plt %*% x))
dimnames(temp) <- list(years=years, mc=seq(1,dim(temp)[2]), species=species_levels)
# Flatten temp
temp <- reshape2::melt(temp)

temp2 <- temp %>% filter(species != "INST") %>% group_by(species, years) %>% summarize(m=mean(value), lo=quantile(value, .05), hi=quantile(value,.95))

ggplot(temp2, aes(x=years)) + geom_ribbon(aes(ymin=lo, ymax=hi), fill='grey50') + geom_line(aes(y=m)) + facet_wrap(~species) + coord_cartesian(ylim=c(-1,1)) + geom_abline(intercept=0, slope=0)
```


```{r}
theta.df <- data.frame(theta1=mc$theta[,1], theta2=mc$theta[,2]) %>% 
  mutate(logtau=(log(tau0)+theta1), logkappa=(log(kappa0)+theta2)) %>%
  mutate(range= sqrt(8)/exp(logkappa), sigma=1/(sqrt(4*pi)*exp(logtau+logkappa))) %>%
  dplyr::select(range, sigma) 
```

```{r}
B.lat.plt <- vector('list', length(species_levels))
names(B.lat.plt) <- species_levels
for(s in species_levels[-1]){ 
    these.trees <- measure %>% filter(species==s)
    r <- range(these.trees$lat)
    B.lat.plt[[s]] <- bs(x=seq(min(r), max(r), length=50), 
                     knots=seq(r[1],r[2], length=k.lat)[-c(1,k.lat)],
                     df=3,
                     Boundary.knots=r, intercept=TRUE)
    dimnames(B.lat.plt[[s]])[[2]] <- paste0('l', 1:(k.lat+3-1))
    # cooerce to sparse
    #attr(B.lat.plt[[s]], 'class') <- 'matrix'
    attr(B.lat.plt[[s]], 'x') <- seq(min(r), max(r), length=50)
 
}

beta.temp <- sweep(mc$beta.lat.1, c(1,2), mc$beta.1, FUN='+')

s <-10
temp <- beta.temp[keeps,s,] %*% t(B.lat.plt[[s]])
dimnames(temp) = list(mc=seq(1:nrow(temp)), lat=seq(1:50))
# Flatten temp
temp <- reshape2::melt(temp)
temp2 <- temp %>% group_by(lat) %>% summarize(m=mean(value), lo=quantile(value, .05), hi=quantile(value,.95))
temp2$lat <- attr(B.lat.plt[[s]], 'x')[unclass(temp2$lat)]

ggplot(temp2, aes(x=lat)) + geom_ribbon(aes(ymin=lo, ymax=hi), fill='grey50') + geom_line(aes(y=m)) + coord_cartesian(ylim=c(-1,1)) + geom_abline(intercept=0, slope=0) +
  geom_rug(data=tree_loc %>% filter(species==species_levels[s]), aes(x=lat))
```
